review nits
